Diabetes Dataset

A diabetes dataset on 1,000 patients.

Usage

diabetes

Format

A data frame with 1,000 rows and 14 variables:

ID

ID.

No_Pation

No. of Patient.

Gender

Gender.

AGE

Age.

Urea

Urea.

Cr

Creatinine ratio(Cr).

HbA1c

HbA1c.

Chol

Cholesterol (Chol).

TG

Triglycerides(TG).

HDL

HDL.

LDL

LDL.

VLDL

VLDL.

BMI

Body Mass Index (BMI).

CLASS

Class (the patient's diabetes disease class may be Diabetic, Non-Diabetic, or Predict-Diabetic).

Source

Rashid A (2020), “Diabetes Dataset”, Mendeley Data, V1, doi: 10.17632/wj9rwkp9c2.1.

Details

The data were collected from the Iraqi society, as they data were acquired from the laboratory of Medical City Hospital and (the Specializes Center for Endocrinology and Diabetes-Al-Kindy Teaching Hospital).

See also

Examples

if (FALSE) { # \dontrun{
data(diabetes)
head(diabetes)
} # }